pyproject.toml
[project]
name = "partsbox-mcp"
version = "0.1.0"
description = "PartsBox MCP Server for parts inventory management"
dependencies = [
    "fastmcp",
    "requests",
    "python-dotenv",
    "jmespath",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpserver>=1.0",
    "responses>=0.25",
]

[project.scripts]
partsbox-mcp = "partsbox_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/partsbox_mcp"]
